EXPRM: The E experiment
Merges `Enter` and `Space` into a single key, the former triggering on
double-tap. Also moves the `Backspace` key to the old `Enter` position,
and replaces the old `Backspace` with `E`.

This addresses #16, but is so very experimental, that I'm not even
putting it onto the master branch.

Requires a QMK with tap-dance functionality, which is only on my own
branch for now.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>